Refactor dns pkg into subpackage
