--- pkg/dns-registrar/.gitignore
@@ -1,4 +1,0 @@
-*.key
-*.pub
-*.sum
-.vscode
