+++ pkg/dns-registrar/encryption/encryption.go
