--- pkg/dns-registrar/go.mod
@@ -1,15 +1,0 @@
-module github.com/mingyech/conjure-dns-registrar
-
-go 1.18
-
-require (
-	github.com/flynn/noise v1.0.0
-	github.com/refraction-networking/utls v1.0.0
-	golang.org/x/crypto v0.0.0-20210322153248-0c34fe9e7dc2
-	golang.org/x/net v0.0.0-20220531201128-c960675eff93
-)
-
-require (
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
-	golang.org/x/text v0.3.7 // indirect
-)
