+++ pkg/dns-registrar/queuepacketconn/consts.go
